freelancer-toolkit/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
playwright
beautifulsoup4
spacy
elevenlabs
PerplexiPy
